fix: create redis store in store module
