+++ store/Cargo.toml
@@ -10,3 +10,6 @@ sqlx = { version = "0.8.6", features = ["runtime-tokio-rustls", "postgres", "uui
 bcrypt = "0.15"
 tokio = { version = "1.0", features = ["full"] }
 tracing = "0.1"
+redis = { version = "0.32.5", features = ["tokio-comp"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
